refactor types to a separate directory
