+++ frontend/src/metabase/embedding/embedding-hub/components/EmbeddingChecklist/index.ts
@@ -1,2 +1,1 @@
 export { EmbeddingChecklist } from "./EmbeddingChecklist";
-export type { EmbeddingStep, EmbeddingStepId } from "./EmbeddingChecklist";
