+++ frontend/src/metabase/embedding/embedding-hub/components/index.ts
@@ -1,6 +1,2 @@
 export { EmbeddingHub } from "./EmbeddingHub";
-export {
-  EmbeddingChecklist,
-  type EmbeddingStep,
-  type EmbeddingStepId,
-} from "./EmbeddingChecklist";
+export { EmbeddingChecklist } from "./EmbeddingChecklist";
