+++ frontend/src/metabase/embedding/embedding-hub/index.ts
@@ -1,3 +1,4 @@
 export * from "./components";
 export * from "./utils";
 export * from "./hooks";
+export * from "./types";
