+++ frontend/src/metabase/embedding/embedding-hub/types/index.ts
@@ -1,0 +1,1 @@
+export type { EmbeddingStep, EmbeddingStepId } from "./embedding-checklist";
